+++ src/main/java/net/soulsweaponry/entity/projectile/noclip/AirCombustion.java
@@ -1,20 +1,14 @@
 package net.soulsweaponry.entity.projectile.noclip;
 
-import net.minecraft.entity.EntityDimensions;
-import net.minecraft.entity.EntityPose;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
-import net.minecraft.entity.data.DataTracker;
-import net.minecraft.entity.data.TrackedData;
-import net.minecraft.entity.data.TrackedDataHandlerRegistry;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.network.packet.s2c.play.EntitySpawnS2CPacket;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundEvents;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import net.soulsweaponry.entity.mobs.DayStalker;
@@ -27,7 +21,6 @@ import net.soulsweaponry.registry.EntityRegistry;
 public class AirCombustion extends DamagingWarmupEntity {
 
     private boolean isEmpowered;
-    private static final TrackedData<Float> RADIUS = DataTracker.registerData(AirCombustion.class, TrackedDataHandlerRegistry.FLOAT);
 
     public AirCombustion(EntityType<? extends PersistentProjectileEntity> entityType, World world) {
         super(entityType, world);
@@ -81,6 +74,7 @@ public class AirCombustion extends DamagingWarmupEntity {
 
     @Override
     public void onTrigger() {
+        super.onTrigger();
         ParticleEvents.airCombustionEvent(this.getWorld(), this.getX(), this.getBodyY(0.5f), this.getZ());
     }
 
@@ -98,53 +92,11 @@ public class AirCombustion extends DamagingWarmupEntity {
         if (nbt.contains("isEmpowered")) {
             this.setEmpowered(nbt.getBoolean("isEmpowered"));
         }
-        if (nbt.contains("Radius")) {
-            this.setRadius(nbt.getFloat("Radius"));
-        }
     }
 
     @Override
     public void writeCustomDataToNbt(NbtCompound nbt) {
         super.writeCustomDataToNbt(nbt);
         nbt.putBoolean("isEmpowered", this.isEmpowered);
-        nbt.putFloat("Radius", this.getRadius());
-    }
-
-    @Override
-    protected void initDataTracker() {
-        super.initDataTracker();
-        this.getDataTracker().startTracking(RADIUS, 1.0F);
-    }
-
-    public void setRadius(float radius) {
-        if (!this.getWorld().isClient) {
-            this.getDataTracker().set(RADIUS, MathHelper.clamp(radius, 0.0F, 32.0F));
-        }
-    }
-
-    @Override
-    public void calculateDimensions() {
-        double d = this.getX();
-        double e = this.getY();
-        double f = this.getZ();
-        super.calculateDimensions();
-        this.setPosition(d, e, f);
-    }
-
-    public float getRadius() {
-        return this.getDataTracker().get(RADIUS);
-    }
-
-    @Override
-    public void onTrackedDataSet(TrackedData<?> data) {
-        if (RADIUS.equals(data)) {
-            this.calculateDimensions();
-        }
-        super.onTrackedDataSet(data);
-    }
-
-    @Override
-    public EntityDimensions getDimensions(EntityPose pose) {
-        return EntityDimensions.changing(this.getRadius(), this.getRadius());
     }
 }