+++ src/main/java/net/soulsweaponry/entity/projectile/noclip/WarmupLightningEntity.java
@@ -1,7 +1,6 @@
 package net.soulsweaponry.entity.projectile.noclip;
 
 import net.minecraft.entity.EntityType;
-import net.minecraft.entity.LightningEntity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.world.World;
@@ -11,6 +10,7 @@ public class WarmupLightningEntity extends DamagingWarmupEntity {
     public WarmupLightningEntity(EntityType<? extends PersistentProjectileEntity> entityType, World world) {
         super(entityType, world);
         this.setDamage(0);
+        this.setEventId(DamagingWarmupEntityEvents.SPAWN_LIGHTNING);
     }
 
     @Override
@@ -21,14 +21,5 @@ public class WarmupLightningEntity extends DamagingWarmupEntity {
     @Override
     public void applyDamageEffects(boolean wasHit, LivingEntity target) {
 
-    }
-
-    @Override
-    public void onTrigger() {
-        if (this.getWorld().isSkyVisible(this.getBlockPos())) {
-            LightningEntity entity = new LightningEntity(EntityType.LIGHTNING_BOLT, this.getWorld());
-            entity.setPos(this.getX(), this.getY(), this.getZ());
-            this.getWorld().spawnEntity(entity);
-        }
     }
 }