+++ src/main/java/net/soulsweaponry/util/CustomDamageSource.java
@@ -19,6 +19,7 @@ public class CustomDamageSource {
     public static RegistryKey<DamageType> BEAM = createType("beam");
     public static RegistryKey<DamageType> DRAGON_MIST = createType("dragon_mist");
     public static RegistryKey<DamageType> PLAYER_LIGHTNING = createType("player_lightning");
+    public static RegistryKey<DamageType> PLAYER_FIRE = createType("player_fire");
 
     public static DamageSource create(World world, RegistryKey<DamageType> key) {
         return new DamageSource(world.getRegistryManager().get(RegistryKeys.DAMAGE_TYPE).getEntry(key).get());